feat: actually show a gradient to the screen
This commit rewrites `Win32ResizeDIBSection` so that it now allocates
one huge block of memory equal to W * H * 4 (bytes per pixel). This
allows us to have a place to write our pixels, and it acts as a buffer.

Additionally, we can now use this buffer to write a gradient. Also,
we're now updating the entire window, not just a subset of it.

Since we're kind, we're also now freeing the bitmap memory when we exit
the program, as well as when we're resizing. This is because the pixels
are invalidated anyway, we have to recalculate everything, so we might
as well free it and start from scratch.

Signed-off-by: Ștefan-Iulian Alecu <165364995+pascalecu@users.noreply.github.com>