feat: now peeking (messages) under the curtain
This commit switches the logic from exiting if GetMessage returns zero
(i.e. a WM_QUIT is being sent) into using PeekMessage, which allows us
to handle the messages ourselves. We're big boys now. Its only use case
right now is to set Running to False whenever we have a WM_QUIT, but we
could use this to do different actions based on the received message
from the queue.

Signed-off-by: Ștefan-Iulian Alecu <165364995+pascalecu@users.noreply.github.com>