feat: add memory funcs + improve resizing
This commit introduces two functions: AllocateMemory and FreeMemory.
They do what it says on the tin. They handle the zero/nil case
gracefully, but if the program can't allocate memory, it just aborts.

Also, this function can now resize itself only if needed (if the width
or the height are different from the current width/height). We also safe
guard against negative width/height or zero (if that's even possible).

TODO: maybe add a default param for the allocation type and the
protect mode. Although at that point it would be VirtualAlloc with extra
steps and guard rails. I'll see.

Signed-off-by: Ștefan-Iulian Alecu <165364995+pascalecu@users.noreply.github.com>