feat: resize DIB section only if actually needed
This commit adds a check to `Win32ResizeDIBSection` that checks if it
needs to be resized iff the width or the height are different. This
should make sense, as resizing implies reallocating the memory, and we
can keep using the same memory.

TODO: check if it's a check we actually need, as we call
`Win32ResizeDIBSection` only within WM_SIZE. The check on
`Win32UpdateWindow` is more useful, as that will be actually called
often, so we only need to resize if necessary. I am not sure if I even
need to do this, as the resize would happen on WM_SIZE, but I believe it
should still be needed on painting the screen.

Signed-off-by: Ștefan-Iulian Alecu <165364995+pascalecu@users.noreply.github.com>